+++ bidsificator/workers/ImportBidsFilesWorker.py
@@ -3,8 +3,8 @@ import multiprocessing as mp
 
 from PyQt6.QtCore import QThread, pyqtSignal
 
-from .BidsFilesProcss import processBidsFiles
-from .protocol import PROGRESS_DONE, PROGRESS_ERROR
+from .BidsFilesProcess import processBidsFiles
+from .import_processor import PROGRESS_DONE, PROGRESS_ERROR
 
 logger = logging.getLogger(__name__)
 
@@ -20,17 +20,13 @@ class ImportBidsFilesWorker(QThread):
         self.subject_name = subject_name
         self.file_list = file_list
         self.contact_labeling_file = contact_labeling_file
-        self.anatomical_modalities = {
-            "T1w (anat)", "T2w (anat)", "T1rho (anat)",
-            "T2* (anat)", "FLAIR (anat)", "CT (anat)"
-        }
 
     def run(self):
         parent_conn, child_conn = mp.Pipe()
         process = mp.Process(
             target=processBidsFiles,
             args=(child_conn, self.dataset_path, self.subject_name, self.file_list,
-                  self.anatomical_modalities, self.contact_labeling_file),
+                  self.contact_labeling_file),
         )
         process.start()